requirements.txt
Adafruit-Blinka==8.66.2
Adafruit-Blinka-Raspberry-Pi5-Neopixel==1.0.0rc2
adafruit-circuitpython-ahtx0==1.0.28
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-ens160==1.0.13
adafruit-circuitpython-framebuf==1.6.10
adafruit-circuitpython-register==1.11.0
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-ssd1306==2.12.22
adafruit-circuitpython-typing==1.12.2
Adafruit-PlatformDetect==3.84.1
Adafruit-PureIO==1.1.11
annotated-doc==0.0.2
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
bidict==0.23.1
binho-host-adapter==0.1.6
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
circuitpython-bmp180==0.2.0
circuitpython-qmc5883l==0.1.3
circup==2.2.3
click==8.3.0
colorzero==2.0
contourpy==1.3.3
cycler==0.12.1
ens160==0.0.2
fastapi==0.120.0
Flask==3.1.2
Flask-SocketIO==5.5.1
fonttools==4.60.1
getrpimodel==0.1.26
gpiozero==2.0.1
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
lgpio==0.2.2.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mh_z19==3.1.7
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pydantic==2.12.3
pydantic_core==2.41.4
pyftdi==0.57.1
pyparsing==3.2.5
pyserial==3.5
python-dateutil==2.9.0.post0
python-engineio==4.12.3
python-socketio==5.14.2
pytz==2025.2
pyusb==1.3.1
requests==2.32.5
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
semver==3.0.4
simple-websocket==1.1.0
six==1.17.0
smbus2==0.5.0
sniffio==1.3.1
starlette==0.48.0
sysv-ipc==1.1.0
toml==0.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
update-checker==0.18.0
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.3
wsproto==1.2.0
